created a new upload component to handle file upload
