+++ src/Components/Connections/style.css
@@ -24,7 +24,6 @@ body {
 
 .content {
   margin: 2rem 1rem;
-  overflow: auto;
 }
 
 .layout {
