+++ src/Components/Connections/table.js
@@ -1,9 +1,6 @@
 import { Table } from 'antd';
-import { UploadOutlined } from '@ant-design/icons';
-import { Button, Upload } from 'antd';
-import { getDoc, doc, updateDoc } from 'firebase/firestore';
+import { getDoc, doc } from 'firebase/firestore';
 import React, { useEffect } from 'react';
-import Papa from 'papaparse';
 import Column from 'antd/es/table/Column';
 import { useAuth0 } from "@auth0/auth0-react";
 import './style.css'
@@ -19,38 +16,6 @@ const PeopleTable = () => {
 
   let connectionsData = useSelector(state => state.connections.connections);
   const dispatch = useDispatch();
-
-
-  const handleFile = (info) => {
-
-    console.log(info);
-    Papa.parse(info.file, {
-      header: true,
-      complete: (result) => {
-        console.log(result.data);
-        dispatch(setConnectionsData(result.data))
-        console.log('papa parse connections ', connectionsData)
-      }
-    });
-    console.log('papa parse connections ', connectionsData);
-    addConnectionsData();
-  }
-
-  const addConnectionsData = async () => {
-    console.log(user.email)
-    try {
-      if (connectionsData) {
-        const ref = doc(db, 'users', user.email);
-        await updateDoc(ref, {
-          connections: connectionsData
-        });
-        console.log('added connections data to DB')
-
-      }
-    } catch (e) {
-      console.error('Error adding document: ', e);
-    }
-  }
 
   const readConnectionsData = async () => {
 
@@ -78,11 +43,13 @@ const PeopleTable = () => {
       readConnectionsData();
     }
   }, [isAuthenticated])
+
   return (
     <>
 
-      <Table pagination={{ pageSize: '5' }}
-        defaultPageSize={5}
+      <Table pagination={{ pageSize: '10' }}
+        defaultPageSize={10}
+        style={{ height: '100%' }}
         dataSource={connectionsData}>
         <Column title='First Name' dataIndex='First Name' key={Math.random()} />
         <Column title='Last Name' dataIndex='Last Name' key={Math.random()} />
@@ -92,9 +59,7 @@ const PeopleTable = () => {
         <Column title='Connected On' dataIndex='Connected On' key={Math.random()} />
       </Table>
 
-      <Upload customRequest={handleFile}>
-        <Button className='uploadCSVButton' icon={<UploadOutlined />}>Upload CSV</Button>
-      </Upload>
+
 
     </>
   );
