+++ src/Components/Jobs/jobCard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Button, Card, Empty, Space } from "antd";
+import { Button, Card } from "antd";
 import { useSelector } from "react-redux";
 import { Typography } from "antd";
 const { Title, Text, Paragraph } = Typography;
@@ -16,10 +16,10 @@ const JobCard = () => {
             ]}
         >
             <Title level={2} >{selectedJob.title}</Title>
-            <Text strong style={{fontSize: '1.2rem'}}> Company: {selectedJob.company}</Text>
-            <Text strong style={{fontSize: '1.2rem', display: 'block', paddingBottom: '1rem'}}> Location: {selectedJob.location}</Text>
+            <Text strong style={{ fontSize: '1.2rem' }}> Company: {selectedJob.company}</Text>
+            <Text strong style={{ fontSize: '1.2rem', display: 'block', paddingBottom: '1rem' }}> Location: {selectedJob.location}</Text>
 
-            <Paragraph style={{fontSize: '1rem'}}>{selectedJob.details}</Paragraph>
+            <Paragraph style={{ fontSize: '1rem' }}>{selectedJob.details}</Paragraph>
 
         </Card>
